+++ library/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/rum/internal/domain/scope/RumViewManagerScope.kt
@@ -20,6 +20,7 @@ import com.datadog.android.rum.internal.vitals.VitalMonitor
 import com.datadog.android.v2.api.InternalLogger
 import com.datadog.android.v2.core.InternalSdkCore
 import com.datadog.android.v2.core.storage.DataWriter
+import java.lang.ref.WeakReference
 import java.util.concurrent.TimeUnit
 
 internal class RumViewManagerScope(
@@ -27,21 +28,24 @@ internal class RumViewManagerScope(
     private val sdkCore: InternalSdkCore,
     private val backgroundTrackingEnabled: Boolean,
     private val trackFrustrations: Boolean,
+    private val viewChangedListener: RumViewChangedListener?,
     internal val firstPartyHostHeaderTypeResolver: FirstPartyHostHeaderTypeResolver,
     private val cpuVitalMonitor: VitalMonitor,
     private val memoryVitalMonitor: VitalMonitor,
     private val frameRateVitalMonitor: VitalMonitor,
-    private val appStartTimeProvider: AppStartTimeProvider = DefaultAppStartTimeProvider()
+    private val appStartTimeProvider: AppStartTimeProvider = DefaultAppStartTimeProvider(),
+    internal var applicationDisplayed: Boolean
 ) : RumScope {
 
     internal val childrenScopes = mutableListOf<RumScope>()
-    internal var applicationDisplayed = false
+    internal var stopped = false
 
     // region RumScope
 
     @WorkerThread
-    override fun handleEvent(event: RumRawEvent, writer: DataWriter<Any>): RumScope {
-        if (!applicationDisplayed) {
+    override fun handleEvent(event: RumRawEvent, writer: DataWriter<Any>): RumScope? {
+        val canDisplayApplication = !stopped && event !is RumRawEvent.StopSession
+        if (!applicationDisplayed && canDisplayApplication) {
             val processImportance = DdRumContentProvider.processImportance
             val isForegroundProcess = processImportance ==
                 ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND
@@ -52,13 +56,19 @@ internal class RumViewManagerScope(
 
         delegateToChildren(event, writer)
 
-        if (event is RumRawEvent.StartView) {
+        if (event is RumRawEvent.StartView && !stopped) {
             startForegroundView(event)
+        } else if (event is RumRawEvent.StopSession) {
+            stopped = true
         } else if (childrenScopes.count { it.isActive() } == 0) {
             handleOrphanEvent(event, writer)
         }
 
-        return this
+        return if (isViewManagerComplete()) {
+            null
+        } else {
+            this
+        }
     }
 
     override fun getRumContext(): RumContext {
@@ -66,12 +76,16 @@ internal class RumViewManagerScope(
     }
 
     override fun isActive(): Boolean {
-        return true
+        return !stopped
     }
 
     // endregion
 
     // region Internal
+
+    fun isViewManagerComplete(): Boolean {
+        return stopped && childrenScopes.isEmpty()
+    }
 
     @WorkerThread
     private fun startApplicationLaunchView(event: RumRawEvent, writer: DataWriter<Any>) {
@@ -139,6 +153,7 @@ internal class RumViewManagerScope(
             this,
             sdkCore,
             event,
+            viewChangedListener,
             firstPartyHostHeaderTypeResolver,
             cpuVitalMonitor,
             memoryVitalMonitor,
@@ -147,6 +162,14 @@ internal class RumViewManagerScope(
         )
         applicationDisplayed = true
         childrenScopes.add(viewScope)
+        viewChangedListener?.onViewChanged(
+            RumViewInfo(
+                keyRef = WeakReference(event.key),
+                name = event.name,
+                attributes = event.attributes,
+                isActive = true
+            )
+        )
     }
 
     @WorkerThread
@@ -185,6 +208,7 @@ internal class RumViewManagerScope(
             RUM_BACKGROUND_VIEW_NAME,
             event.eventTime,
             emptyMap(),
+            viewChangedListener,
             firstPartyHostHeaderTypeResolver,
             NoOpVitalMonitor(),
             NoOpVitalMonitor(),
@@ -202,6 +226,7 @@ internal class RumViewManagerScope(
             RUM_APP_LAUNCH_VIEW_NAME,
             time,
             emptyMap(),
+            viewChangedListener,
             firstPartyHostHeaderTypeResolver,
             NoOpVitalMonitor(),
             NoOpVitalMonitor(),