+++ library/dd-sdk-android-rum/src/test/kotlin/com/datadog/android/rum/internal/monitor/DatadogRumMonitorTest.kt
@@ -1298,7 +1298,7 @@ internal class DatadogRumMonitorTest {
         val otherScopes = forge.aList {
             forge.anElementFrom(mock(), mock<RumActionScope>(), mock<RumResourceScope>())
         }
-        whenever(mockRumApplicationScope.childScope) doReturn mockSessionScope
+        whenever(mockRumApplicationScope.activeSession) doReturn mockSessionScope
         whenever(mockSessionScope.childScope) doReturn mockViewManagerScope
         whenever(mockViewManagerScope.childrenScopes)
             .thenReturn((viewScopes + otherScopes).toMutableList())
@@ -1334,7 +1334,7 @@ internal class DatadogRumMonitorTest {
         val otherScopes = forge.aList {
             forge.anElementFrom(mock(), mock<RumActionScope>(), mock<RumResourceScope>())
         }
-        whenever(mockRumApplicationScope.childScope) doReturn mockSessionScope
+        whenever(mockRumApplicationScope.activeSession) doReturn mockSessionScope
         whenever(mockSessionScope.childScope) doReturn mockViewManagerScope
         whenever(mockViewManagerScope.childrenScopes)
             .thenReturn((viewScopes + otherScopes).toMutableList())
@@ -1357,7 +1357,7 @@ internal class DatadogRumMonitorTest {
         val listener = mock<RumDebugListener>()
         testedMonitor.debugListener = listener
 
-        whenever(mockRumApplicationScope.childScope) doReturn forge.anElementFrom(
+        whenever(mockRumApplicationScope.activeSession) doReturn forge.anElementFrom(
             mock(),
             mock<RumViewScope>(),
             mock<RumActionScope>(),